backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
openai
python-dotenv
pymupdf
python-docx
faiss-cpu
python-multipart
pydantic 